requirements.txt
fastapi
sqlalchemy
argon2-cffi
pydantic
aiofiles
pillow